refactor: move invoice line mapper
